refactor #22 styled-component 미변경이름 변경
