+++ src/app.jsx
@@ -9,7 +9,7 @@ import MessagePage from "./pages/message-list";
 import RecipientPostPage from "./pages/recipient-post-page";
 import SendMessagePage from "./pages/send-message-page";
 import TestPage from "./pages/test-page";
-import PostPage from "./pages/post-page";
+import PostPage from "./pages/create-post-page";
 import SendPage from "./pages/send-page";
 
 function Provider({ children }) {
@@ -69,7 +69,6 @@ function App() {
           </Route>
           <Route path="/test-components" element={<TestPage />} />
           <Route path="/list" element={<MessagePage />} />
-          <Route path="/post" element={<PostPage />} />
           <Route path="/post/{id}/message" element={<SendPage />} />
         </Routes>
       </BrowserRouter>
